% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metGeneratoR.R
\docType{package}
\name{metGeneratoR}
\alias{metGeneratoR}
\title{metGeneratoR package}
\description{
The metGeneratoR package disaggregates daily NetCDF data to subdaily data.
This can be 1hourly, 2hourly, 3hourly, 4hourly, 6hourly or 8hourly.
If needed it converts data to the output unit needed for VIC.
}
\examples{
########################
## preprocessing
## TODO
########################

########################
## do the job (END)  
## 
## The example below 
########################

## (re-)Initialize the metGen environment with basic settings
##
## Please always rerun this function before making adjustments in the metGeneratoR setup
mgsetInit()

## Set the period to run
mgsetPeriod(startdate = "1950-1-1", enddate = "1950-1-05")

## Set the output timesteps in hours
mgsetOutDt(6)  # 6 hourly output

## Set the input variables
mgsetInVars(list(
  pr         = list(ncname = "pr",      filename = metGen$internal$ncFileNamePr),
  tasmin     = list(ncname = "tasmin",  filename = metGen$internal$ncFileNameTasmin),
  tasmax     = list(ncname = "tasmax",  filename = metGen$internal$ncFileNameTasmax),
  wind       = list(ncname = "sfcWind", filename = metGen$internal$ncFileNameWind),
  swdown     = list(ncname = "rsds",    filename = metGen$internal$ncFileNameswdown),
  lwdown     = list(ncname = "rlds",    filename = metGen$internal$ncFileNamelwdown),
  relhum     = list(ncname = "hurs",    filename = metGen$internal$ncFileNameRelhum),
  psurf      = list(ncname = "ps",      filename = metGen$internal$ncFileNamePs)
))

## Set the output variables
mgsetOutVars(c("tas", "pr", "wind", "vp", "psurf", "swdown", "lwdown", "radfrac"))

## Run the metGeneratoR based on the settings above
## The output files will be written to (a subfolder of) the current working directory of R
metGenRun()
}
\keyword{internal}
