% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metGenRunDemo.R
\name{metGenRunDemo}
\alias{metGenRunDemo}
\title{Run metGeneratoR demo}
\usage{
metGenRunDemo()
}
\value{
This function will produce some information on the screen and if everything succeeds,
it will also produce NetCDF files in the standard output directory
}
\description{
This function runs a demo setup of metGeneratoR.
It uses example NetCDF data as input which comes with this package. It disaggregates this
data and it generates some NetCDF files as output.
This function is mainly usefull to test if the package is really working.
}
\examples{
metGenRunDemo()
}
