% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metGenSettings.R
\name{mgsetInDt}
\alias{mgsetInDt}
\title{Set the input timesteps in n-hours per day}
\usage{
mgsetInDt(inDt)
}
\description{
This function is optional. If not given metGen assumes daily input
}
\examples{
## 24 hourly (daily) input
mgsetInDt(24)

## 6 hourly input
mgsetInDt(6)
}
