% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metGenSettings.R
\name{mgsetInit}
\alias{mgsetInit}
\title{Initialize and set some settings required for the metGen package to run}
\usage{
mgsetInit()
}
\value{
nothing dfgdg
}
\description{
Initialize and set some settings required for the metGen package to run
}
\examples{
mgsetInit()
}
