% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metGenSettings.R
\name{mgsetOutName}
\alias{mgsetOutName}
\title{Set the output filename(s) structure and path(s)}
\usage{
mgsetOutName(nameString, message = F)
}
\description{
if not given, the output filename will have the following structure
"<VAR>\emph{<SYEAR><SMONTH><SDAY>}<EYEAR><EMONTH><EDAY>.nc"
#and will be written to the current working directory
#eg: lwdown_19500101_19501231.nc
}
\details{
#subsitutions/wildcards/aliases
}
\examples{
mgsetOutName("forcing_<VAR>_<SYEAR>.nc", message = TRUE)

mgsetOutName("<VAR>_<SYEAR><SMONTH><SDAY>.nc", message = TRUE)

mgsetOutName("<VAR>_<SYEAR><SMONTH><SDAY>_<EYEAR><EMONTH><EDAY>.nc", message = TRUE)

mgsetOutName("output/<VAR>/<VAR>_<SYEAR><SMONTH><SDAY>_<EYEAR><EMONTH><EDAY>.nc", message = TRUE)

}
