% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metGenSettings.R
\name{mgsetXYbox}
\alias{mgsetXYbox}
\title{Set the XYbox}
\usage{
mgsetXYbox(startX, endX, startY, endY)
}
\description{
Use this function in case you want to select a subset of the original input
}
\examples{
mgsetXYbox(startX = 330, endX = 450, startY = 30, endY = 130)

}
