% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metGeneratoR.R
\docType{package}
\name{metGeneratoR}
\alias{metGeneratoR}
\title{metGeneratoR package}
\description{
The metGeneratoR package disaggregates daily NetCDF data to subdaily data.
This can be hourly, 3hourly, 6hourly, etc.
If needed it converts data to the proper output unit.
}
\examples{



########################
## preprocessing
########################

########################
## do the job (END)  
## 
## The example below 
########################

## (re-)Initialize the metGen environment with basic settings
##
## Please always rerun this function before making adjustments in the metGeneratoR setup
mgsetInit()

## Set the XYbox (optional)
## 
## Use this function in case you want to select a subset of the original input
## usage: mgsetXYbox(startX, endX, startY, endY)
mgsetXYbox(330, 450, 30, 130)

## Set the period to run
mgsetPeriod(startdate = "1950-1-1", enddate = "1950-1-31")

## Set the output timesteps in hours
mgsetInDt(24)  # 24 hourly input
mgsetOutDt(6)  # 6 hourly output

## Set the input variables
mgsetInVars(list(
  pr         = list(ncname = "pr",      filename = metGen$internal$ncFileNamePr),
  tasmin     = list(ncname = "tasmin",  filename = metGen$internal$ncFileNameTasmin),
  tasmax     = list(ncname = "tasmax",  filename = metGen$internal$ncFileNameTasmax),
  wind       = list(ncname = "sfcWind", filename = metGen$internal$ncFileNameWind),
  swdown     = list(ncname = "rsds",    filename = metGen$internal$ncFileNameswdown),
  lwdown     = list(ncname = "rlds",    filename = metGen$internal$ncFileNamelwdown),
  relhum     = list(ncname = "hurs",    filename = metGen$internal$ncFileNameRelhum),
  psurf      = list(ncname = "ps",      filename = metGen$internal$ncFileNamePs)
))

## Set the output variables
mgsetOutVars(c("tas", "pr", "wind", "vp", "psurf", "swdown", "lwdown", "radfrac"))

## Set the output filename(s) structure and path(s) 
mgsetOutName("output/<VAR>_<SYEAR><SMONTH><SDAY>_<EYEAR><EMONTH><EDAY>.nc", message = TRUE)

## Retreive dimension info from the input files
mggetInDims()

## Set dimension info for the output files
mgsetOutDims()

## Check the input variables
mgcheckInVars()

## Check the output variables
mgcheckOutVars()

## Run the metGeneratoR based on the settings above
metGenRun()
}
\keyword{internal}
