% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metGeneratoR.R
\docType{package}
\name{metGeneratoR}
\alias{metGeneratoR}
\alias{metGeneratoR-package}
\title{metGeneratoR package}
\description{
The metGeneratoR package disaggregates daily NetCDF data to subdaily data.
This can be hourly, 3hourly, 6hourly, etc.
If needed it converts data to the proper output unit.
}
\examples{

## Set the latlonbox
mgsetLonlatbox(c(92.25, 110.25, 7.25, 36.25))

## Set the period to run
mgsetPeriod(startdate = "1950-1-1", enddate = "1950-1-31")

## Set the output timesteps in hours
mgsetOutDt(6) # 6 hourly output

## Set the input variables
mgsetInVars(list(
  pr         = list(ncname = "pr",      filename = metGen$internal$ncFileNamePr),
  tasmin     = list(ncname = "tasmin",  filename = metGen$internal$ncFileNameTasmin),
  tasmax     = list(ncname = "tasmax",  filename = metGen$internal$ncFileNameTasmax),
  wind       = list(ncname = "wind",    filename = metGen$internal$ncFileNameWind)
))

## Set the output variables
mgsetOutVars(c("tas", "pr", "wind"))

## Set the output filename(s) structure and path(s) 
mgsetOutName("output/<VAR>/<VAR>_<SYEAR><SMONTH><SDAY>_<EYEAR><EMONTH><EDAY>.nc")

## Run the metGeneratoR based on the settings above
metGenRun()
}
\keyword{internal}
